requirements.txt
Django==5.2.9
djangorestframework==3.16.1
drf-yasg==1.21.11
django-cors-headers==4.9.0
pytz==2025.2
sqlparse==0.5.4
uritemplate==4.2.0
inflection==0.5.1
PyYAML==6.0.2
asgiref==3.11.0
typing_extensions==4.14.1
requests==2.32.3
python-dotenv==1.0.1
gunicorn>=22.0
whitenoise[brotli]>=6.6
feedparser>=6.0.0
insightface==0.7.3
opencv-python==4.10.0.84
numpy==1.26.4
fastapi==0.115.0
uvicorn==0.30.1
websockets==12.0
selenium==4.21.0
beautifulsoup4==4.12.3
snscrape==0.7.0.20230622
fake-useragent==1.5.1
requests[socks]==2.32.3
undetected-chromedriver==3.5.5